moved and adjusted WhiteBox tests (test/lib-test/sun/hotspot/whitebox)
