+++ test/lib-test/jdk/test/whitebox/vm_flags/StringTest.java
@@ -22,30 +22,28 @@
  */
 
 /*
- * @test SizeTTest
- * @bug 8054823
+ * @test StringTest
+ * @bug 8028756
  * @library /test/lib
  * @modules java.base/jdk.internal.misc
- *          java.management/sun.management
- * @build sun.hotspot.WhiteBox
- * @run driver jdk.test.lib.helpers.ClassFileInstaller sun.hotspot.WhiteBox
- * @run main/othervm/timeout=600 -Xbootclasspath/a:. -XX:+UnlockDiagnosticVMOptions -XX:+WhiteBoxAPI -XX:+UnlockExperimentalVMOptions SizeTTest
- * @summary testing of WB::set/getSizeTVMFlag()
+ * @modules java.management/sun.management
+ * @build jdk.test.whitebox.WhiteBox
+ * @run driver jdk.test.lib.helpers.ClassFileInstaller jdk.test.whitebox.WhiteBox
+ * @run main/othervm/timeout=600 -Xbootclasspath/a:. -XX:+UnlockDiagnosticVMOptions -XX:+WhiteBoxAPI StringTest
+ * @summary testing of WB::set/getStringVMFlag()
+ * @author igor.ignatyev@oracle.com
  */
-import jdk.test.lib.Platform;
 
-public class SizeTTest {
-    private static final String FLAG_NAME = "ArrayAllocatorMallocLimit";
-    private static final Long[] TESTS = {0L, 100L, (long) Integer.MAX_VALUE,
-        (1L << 32L) - 1L, 1L << 32L};
-    private static final Long[] EXPECTED_64 = TESTS;
-    private static final Long[] EXPECTED_32 = {0L, 100L,
-        (long) Integer.MAX_VALUE, (1L << 32L) - 1L, 0L};
+public class StringTest {
+    private static final String FLAG_NAME = "CompileOnly";
+    private static final String FLAG_DEBUG_NAME = "SuppressErrorAt";
+    private static final String[] TESTS = {"StringTest::*", ""};
 
     public static void main(String[] args) throws Exception {
         VmFlagTest.runTest(FLAG_NAME, TESTS,
-            Platform.is64bit() ? EXPECTED_64 : EXPECTED_32,
-            VmFlagTest.WHITE_BOX::setSizeTVMFlag,
-            VmFlagTest.WHITE_BOX::getSizeTVMFlag);
+            VmFlagTest.WHITE_BOX::setStringVMFlag,
+            VmFlagTest.WHITE_BOX::getStringVMFlag);
+        VmFlagTest.runTest(FLAG_DEBUG_NAME, VmFlagTest.WHITE_BOX::getStringVMFlag);
     }
 }
+
